ultrasone verbetering
de ultrasone meet nu een stuk preciezer.  de delay tussen het detecteren van een obstakel en het stoppen van de agv is verholpen. PID werkt nog neit

+++ ultrasone.h
@@ -26,10 +26,10 @@
 #define TRIG_PIN PH0
 #define ECHO_PIN PH1
 void init_ultrasonic(void);
-uint16_t measure_distance(void);
+float measure_distance(void);
 			// Return the distance in centimeters
 
 
 void init_ultrasoniclat(void);
-uint16_t measure_distancelat(void);
+float measure_distancelat(void);
 #endif // ULTRASONETEST_H_INCLUDED
